requirements.txt
torch
streamlit
numpy
matplotlib
scikit-learn
google-generativeai
